image_service/requirements.txt
PyJWT==1.5.2
pytz==2017.2
pyasn1-modules==0.0.9
google-cloud-storage==1.13.1